modules and auto refresh for hsl data
