+++ src/modules/coordinate-calc.js
@@ -3,23 +3,24 @@ import {fazerDataFiArabia, fazerDataFiKaramalmi, sodexoDataMyllypuro, sodexoData
 
 
 // distance calculator for coordinates, from the wise people in internet
-function getDistanceFromLatLonInKm(lat1,lon1,lat2,lon2) {
+function getDistanceFromLatLonInKm(lat1, lon1, lat2, lon2) {
   const earthRadius = 6371; // Radius of the earth in km
-  const dLat = deg2rad(lat2-lat1);  // deg2rad below
-  const dLon = deg2rad(lon2-lon1);
+  const dLat = deg2rad(lat2 - lat1);  // deg2rad below
+  const dLon = deg2rad(lon2 - lon1);
   const a =
-    Math.sin(dLat/2) * Math.sin(dLat/2) +
+    Math.sin(dLat / 2) * Math.sin(dLat / 2) +
     Math.cos(deg2rad(lat1)) * Math.cos(deg2rad(lat2)) *
-    Math.sin(dLon/2) * Math.sin(dLon/2)
+    Math.sin(dLon / 2) * Math.sin(dLon / 2)
   ;
-  const c = 2 * Math.atan2(Math.sqrt(a), Math.sqrt(1-a));
+  const c = 2 * Math.atan2(Math.sqrt(a), Math.sqrt(1 - a));
    // Distance in km
   return earthRadius * c;
 }
 
 function deg2rad(deg) {
-  return deg * (Math.PI/180);
+  return deg * (Math.PI / 180);
 }
+
 let nearestRestaurant;
 let nearestCoordinate = null;
 let smallestDistance = Infinity;
@@ -36,7 +37,7 @@ const calcDistance = (lat, lon) => {
       nearestCoordinate = coord;
     }
   }
-nearestRestaurant = nearestCoordinate;
+  nearestRestaurant = nearestCoordinate;
 };
 
 const showNearestRestaurantMenu = () => {
